background move

+++ Blocks.java
@@ -4,7 +4,7 @@ public class Blocks {
     int blockx = (int) (Math.random() * 1000);
     int blocky = (int) (Math.random() * 1000);
 
-    public void allThatIsAboveAndAllThatIsBelow(int vy) {
+    public void allThatIsAboveAndAllThatIsBelow(double vy) {
         this.blocky -= vy;
     }
 }
